backend/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
starlette==0.38.6
uvicorn[standard]==0.32.0
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.3
numpy==1.26.4
shap==0.46.0
fpdf2==2.8.2
httpx==0.28.1
python-dotenv==1.0.1
firebase-admin==6.6.0
pytest==8.3.4
httpx[http2]==0.28.1
